+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ### docs(readme): add README docs
 
 구현할 기능에 대한 명세와 절차를 정리하여 README.md 파일에 작성함
+
+### feat(Application): add default I/O and exception handling
+
+- `Console.readline()` 을 활용한 입력 기능 구현
+- `calculate()` 메서드 생성하고 예외 시 예외 출력 후 프로그램 종료하도록 함